requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
jinja2
jinja2
